refactor : 생성/수정에 필요한 공통 UI를 core:ui로 변경
