+++ feature/content/create/src/commonMain/kotlin/com/whatever/caramel/feature/content/create/ContentCreateScreen.kt
@@ -47,6 +47,7 @@ import com.whatever.caramel.core.designsystem.foundations.Resources
 import com.whatever.caramel.core.designsystem.themes.CaramelTheme
 import com.whatever.caramel.core.domain.entity.Tag
 import com.whatever.caramel.core.ui.content.ContentAssigneeChipRow
+import com.whatever.caramel.core.ui.content.ContentScheduleInfo
 import com.whatever.caramel.core.ui.content.ContentTextArea
 import com.whatever.caramel.core.ui.content.CreateMode
 import com.whatever.caramel.core.ui.content.CreateModeSwitch
@@ -57,7 +58,6 @@ import com.whatever.caramel.core.ui.content.TitleTextField
 import com.whatever.caramel.core.ui.picker.CaramelDatePicker
 import com.whatever.caramel.core.ui.picker.CaramelTimePicker
 import com.whatever.caramel.core.ui.util.rememberKeyboardVisibleState
-import com.whatever.caramel.feature.content.create.component.ContentScheduleInfo
 import com.whatever.caramel.feature.content.create.mvi.ContentCreateIntent
 import com.whatever.caramel.feature.content.create.mvi.ContentCreateState
 import com.whatever.caramel.feature.content.create.mvi.ScheduleDateTimeType
@@ -307,7 +307,7 @@ internal fun ContentScreen(
                             ContentScheduleInfo(
                                 modifier = Modifier.fillMaxWidth(),
                                 leadingText = "시작",
-                                dateTimeInfo = state.startDateTimeInfo.dateTime,
+                                dateTimeInfo = state.startDateTime.dateTime,
                                 onClickDate = {
                                     onIntent(
                                         ContentCreateIntent.ClickDate(
@@ -328,7 +328,7 @@ internal fun ContentScreen(
                             ContentScheduleInfo(
                                 modifier = Modifier.fillMaxWidth(),
                                 leadingText = "종료",
-                                dateTimeInfo = state.endDateTimeInfo.dateTime,
+                                dateTimeInfo = state.endDateTime.dateTime,
                                 onClickDate = {
                                     onIntent(
                                         ContentCreateIntent.ClickDate(
